requirements.txt
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
cryptography==40.0.2
Deprecated==1.2.13
gitdb==4.0.10
GitPython==3.1.31
idna==3.4
pycparser==2.21
PyGithub==1.58.2
PyJWT==2.7.0
PyNaCl==1.5.0
PyYAML==6.0
requests==2.31.0
smmap==5.0.0
urllib3==1.26.6
wrapt==1.15.0
